Updated nuget packages

+++ src/SFA.DAS.EAS.Employer_Financial.Database/Tables/AccountTransferAllowanceSnapshot.sql
@@ -2,7 +2,7 @@
 (
     [Id] INT NOT NULL PRIMARY KEY IDENTITY, 
     [AccountId] BIGINT NOT NULL, 
-    [Year] SMALLINT NOT NULL, 
+    [Year] INT NOT NULL, 
     [TransferAllowance] DECIMAL(18, 5) NOT NULL, 
     [SnapshotTime] DATETIME NOT NULL 
     )
